+++ astropy/modeling/fitting.py
@@ -329,7 +329,9 @@ class LinearLSQFitter(metaclass=_FitterMeta):
             equation has to be solved for each model separately when their
             co-ordinate grids differ.
         weights : array (optional)
-            weights
+            Weights for fitting.
+            For data with Gaussian uncertainties, the weights should be
+            1/sigma.
         rcond :  float, optional
             Cut-off ratio for small singular values of ``a``.
             Singular values are set to zero if they are smaller than ``rcond``
@@ -710,7 +712,9 @@ class LevMarLSQFitter(metaclass=_FitterMeta):
         z : array (optional)
            input coordinates
         weights : array (optional)
-           weights
+            Weights for fitting.
+            For data with Gaussian uncertainties, the weights should be
+            1/sigma.
         maxiter : int
             maximum number of iterations
         acc : float
@@ -850,7 +854,9 @@ class SLSQPLSQFitter(Fitter):
         z : array (optional)
             input coordinates
         weights : array (optional)
-            weights
+            Weights for fitting.
+            For data with Gaussian uncertainties, the weights should be
+            1/sigma.
         kwargs : dict
             optional keyword arguments to be passed to the optimizer or the statistic
 
@@ -919,7 +925,9 @@ class SimplexLSQFitter(Fitter):
         z : array (optional)
             input coordinates
         weights : array (optional)
-            weights
+            Weights for fitting.
+            For data with Gaussian uncertainties, the weights should be
+            1/sigma.
         kwargs : dict
             optional keyword arguments to be passed to the optimizer or the statistic